dualtext_server/requirements.txt
asgiref==3.3.1
astroid==2.4.2
dataclasses==0.6
Django==3.1.3
django-cors-headers==3.5.0
django-elasticsearch-dsl==7.1.4
djangorestframework==3.12.2
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
filelock==3.0.12
future==0.18.2
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.5
numpy==1.19.4
packaging==20.4
protobuf==3.14.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentence-transformers==0.3.9
sentencepiece==0.1.91
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tokenizers==0.9.3
toml==0.10.2
torch==1.7.0
tqdm==4.52.0
transformers==3.5.1
typing-extensions==3.7.4.3
wrapt==1.12.1
